The rest of Aki Koskinen's patches, adding linux support:
1) Added CMakeLists.txt and build.sh to make building on Linux easy. Note that build.sh is supposed to be exacutable (it says so in the patch file (*) but I don't know if that notion goes all the way to the repository). Short build instructions in ReadMe.txt.
2) Put Win specific fopen and vsnprintf variants inside preprocessor guards and use the C standard ones on non-windows systems (windows could probably use the standard ones as well but I left it unchanged since I didn't want to touch the windows version without knowing all the possible details).
3) The way of accessing PVRTextureUtilities has apparently changed in PVRTexLib version 3.20. Added a preprocessor check that should work with both the old and new interface but untested on PVRTexLib versions < 3.20.
4) Added std namespace and some letter case changes in various places.
* adds a cmake build script for linux for the RTBareBones example project. No code changes were needed.
* Also added a small set of instructions to RTBareBones/android/readme.txt on how to build the Android version in Linux. 
* adds a cmake build script for linux for the RTSimpleApp example project. Few code changes made:
1) Use AudioManagerSDL on Linux.
2) If playing of the mp3 fails, try to play the ogg. 
This patch adds a couple of tools that can be used on Linux. They are called buildAndRun.sh and update_media.sh and they are placed in tools/linux directory. Both scripts should be executable (once again, says so in the patch).

The buildAndRun.sh script does what the name says: builds and runs a project. It calls the needed cmake and make commands for building. It also calls the update_media.sh script which is similar to the update_media.bat scripts for Windows.

There aren't any usage instructions for either script anywhere but those could be added to the wiki for example.

To note: update_media.sh (and hence buildAndRun.sh also since it calls update_media.sh) isn't safe for RTBareBones since it erases the bin/interface directory which is svn controlled. This isn't checked or told anywhere